# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for auth with JWT",
        "title": "Auth Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes current access token and all associated refresh tokens. Requires valid JWT in Authorization header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the user ID from valid JWT token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Get current user ID",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.UserIdResponse"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates new access and refresh tokens pair using valid refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh authentication tokens",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New tokens pair",
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid, expired or revoked tokens",
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/token": {
            "post": {
                "description": "Creates new access and refresh tokens pair for specified user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Generate new authentication tokens",
                "parameters": [
                    {
                        "description": "User ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.UserIdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully generated tokens",
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or empty user ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to generate or save tokens",
                        "schema": {
                            "$ref": "#/definitions/github_com_redeflesq_auth-example_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_redeflesq_auth-example_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_redeflesq_auth-example_internal_model.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "string"
                }
            }
        },
        "github_com_redeflesq_auth-example_internal_model.TokenRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "description": "It's not hash",
                    "type": "string"
                }
            }
        },
        "github_com_redeflesq_auth-example_internal_model.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_redeflesq_auth-example_internal_model.UserIdRequest": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_redeflesq_auth-example_internal_model.UserIdResponse": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer eyJhbGciOi...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}